pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]
exclude = ["client*", "tests*", "scripts*", "docs*"]

[project]
name = "databricks-phoenix-app-template"
version = "0.1.0"
description = "Databricks GenAI App Template (Phoenix)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.115.8",
  "mlflow[databricks]>=3.1",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.34.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "requests>=2.32.0",
  "databricks-sdk>=0.44.1",
  # Conflict resolution pins for Databricks Apps pre-installed packages
  "tenacity==9.0.0",
  "pillow==11.1.0",
  "websockets==15.0",
  "pyarrow==18.1.0",
  "markupsafe==3.0.2",
  "rich==14.0.0",
  "protobuf>=3.12.0,<5",
  "pandas==2.3.0",
  "flask==3.1.0",
  "werkzeug==3.1.3",
  "sqlalchemy[asyncio]>=2.0.41",
  "alembic>=1.16.1",
  "asyncpg>=0.30.0",
  "greenlet>=3.0.0",
  "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
  "click>=8.1.8",
  "ruff>=0.9.6",
  "watchdog[watchmedo]>=6.0.0",
  "databricks-connect==16.1.6",
]

[tool.uv]
prerelease = "allow"


[tool.ruff]
src = ["server"]
line-length = 100
indent-width = 2

# Enable Pyflakes `F`, pycodestyle "W" and `E`, "Q" quotes, "I" imports.
fix = true
ignore = [
  "D105", # Missing docstring in magic method.
  "D106", # Missing docstring in a public nested class.
  "D107", # Missing docstring in __init__.
  "D203", # 1 blank line required before class docstring.
  "D204", # 1 blank line required after class docstring.
  "D213", # Multi-line docstring summary should start at the second line.
  "D401", # First line of docstring should be in imperative mood: "{first_line}"
  "D407", # Missing dashed underline after section.
]
select = ["E", "W", "F", "Q", "I", "D"]

# Exclude a variety of commonly ignored directories.
exclude = [
  "__pycache__",
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.flake8-quotes]
inline-quotes = "single"

[tool.ruff.per-file-ignores]
"*_test.py" = ["D101", "D102", "D103", "D104"]
"__init__.py" = ["D104"]

[tool.ruff.pep8-naming]
# Allow Pydantic's `@validator` decorator to trigger class method treatment.
classmethod-decorators = ["classmethod", "pydantic.validator"]

[tool.ruff.lint.pydocstyle]
convention = "google"
